refactor(architecture): fix misplaced responsibilities between application and infrastructure
