+++ src/aigora/curriculum_graph/infrastructure/files/assembling/assembler_errors.py
