+++ src/aigora/curriculum_graph/infrastructure/files/assembling/graph_assembler.py
@@ -1,9 +1,9 @@
 from __future__ import annotations
 
-from aigora.curriculum_graph.domain.curriculum_graph import CurriculumGraph
-from aigora.curriculum_graph.domain.curriculum_profile import CurriculumProfile
-from aigora.curriculum_graph.domain.edge import Edge
-from aigora.curriculum_graph.domain.node import Node
+from aigora.curriculum_graph.domain.entities.curriculum_graph import CurriculumGraph
+from aigora.curriculum_graph.domain.entities.curriculum_profile import CurriculumProfile
+from aigora.curriculum_graph.domain.entities.edge import Edge
+from aigora.curriculum_graph.domain.entities.node import Node
 
 from .assembler_errors import (
     DuplicateNodeError,
