+++ src/aigora/curriculum_graph/infrastructure/files/mapping/mapper_errors.py
