+++ src/aigora/curriculum_graph/infrastructure/files/parsing/graph_parser.py
