+++ src/aigora/curriculum_graph/infrastructure/files/parsing/parser_errors.py
