+++ src/aigora/curriculum_graph/infrastructure/files/serialization/serializer_errors.py
