+++ src/aigora/curriculum_graph/infrastructure/files/validation/schema_errors.py
