+++ tests/unit/curriculum_graph/application/loading/test_graph_loader.py
@@ -2,12 +2,12 @@ import pytest
 
 from aigora.curriculum_graph.application.loading.graph_loader import GraphLoader
 from aigora.curriculum_graph.application.loading.loader_errors import GraphLoaderError
-from aigora.curriculum_graph.domain.curriculum_graph import CurriculumGraph
-from aigora.curriculum_graph.domain.curriculum_profile import CurriculumProfile
-from aigora.curriculum_graph.domain.edge import Edge
-from aigora.curriculum_graph.domain.enums import EdgeType, MasteryLevel
-from aigora.curriculum_graph.domain.mastery import MasteryCriterion, MasteryScale
-from aigora.curriculum_graph.domain.node import Node
+from aigora.curriculum_graph.domain.entities.curriculum_graph import CurriculumGraph
+from aigora.curriculum_graph.domain.entities.curriculum_profile import CurriculumProfile
+from aigora.curriculum_graph.domain.entities.edge import Edge
+from aigora.curriculum_graph.domain.enums.enums import EdgeType, MasteryLevel
+from aigora.curriculum_graph.domain.value_objects.mastery import MasteryCriterion, MasteryScale
+from aigora.curriculum_graph.domain.entities.node import Node
 
 
 def make_mastery_scale() -> MasteryScale:
