+++ tests/unit/curriculum_graph/application/queries/test_graph_query.py
@@ -5,11 +5,11 @@ from aigora.curriculum_graph.application.queries.query_errors import (
     NodeNotFoundError,
     PathNotFoundError,
 )
-from aigora.curriculum_graph.domain.curriculum_graph import CurriculumGraph
-from aigora.curriculum_graph.domain.edge import Edge
-from aigora.curriculum_graph.domain.enums import EdgeType, MasteryLevel
-from aigora.curriculum_graph.domain.mastery import MasteryCriterion, MasteryScale
-from aigora.curriculum_graph.domain.node import Node
+from aigora.curriculum_graph.domain.entities.curriculum_graph import CurriculumGraph
+from aigora.curriculum_graph.domain.entities.edge import Edge
+from aigora.curriculum_graph.domain.enums.enums import EdgeType, MasteryLevel
+from aigora.curriculum_graph.domain.value_objects.mastery import MasteryCriterion, MasteryScale
+from aigora.curriculum_graph.domain.entities.node import Node
 
 
 def make_mastery_scale() -> MasteryScale:
