+++ tests/unit/curriculum_graph/application/serialization/test_graph_serializer.py
@@ -3,16 +3,16 @@ import json
 import pytest
 import yaml
 
-from aigora.curriculum_graph.application.serialization.graph_serializer import GraphSerializer
-from aigora.curriculum_graph.application.serialization.serializer_errors import (
+from aigora.curriculum_graph.infrastructure.files.serialization.graph_serializer import GraphSerializer
+from aigora.curriculum_graph.infrastructure.files.serialization.serializer_errors import (
     UnsupportedSerializationFormatError,
 )
-from aigora.curriculum_graph.domain.curriculum_graph import CurriculumGraph
-from aigora.curriculum_graph.domain.curriculum_profile import CurriculumProfile
-from aigora.curriculum_graph.domain.edge import Edge
-from aigora.curriculum_graph.domain.enums import EdgeType, MasteryLevel
-from aigora.curriculum_graph.domain.mastery import MasteryCriterion, MasteryScale
-from aigora.curriculum_graph.domain.node import Node
+from aigora.curriculum_graph.domain.entities.curriculum_graph import CurriculumGraph
+from aigora.curriculum_graph.domain.entities.curriculum_profile import CurriculumProfile
+from aigora.curriculum_graph.domain.entities.edge import Edge
+from aigora.curriculum_graph.domain.enums.enums import EdgeType, MasteryLevel
+from aigora.curriculum_graph.domain.value_objects.mastery import MasteryCriterion, MasteryScale
+from aigora.curriculum_graph.domain.entities.node import Node
 
 
 # ── Helpers ───────────────────────────────────────────────────────────────────
