+++ tests/unit/curriculum_graph/application/validation/test_graph_persistence_validator.py
@@ -3,17 +3,17 @@ from __future__ import annotations
 
 import pytest
 
-from aigora.curriculum_graph.application.validation.graph_persistence_validator import (
+from aigora.curriculum_graph.infrastructure.neo4j.validation.graph_persistence_validator import (
     GraphPersistenceValidationError,
     GraphPersistenceValidator,
     PersistenceValidationResult,
 )
-from aigora.curriculum_graph.domain.curriculum_graph import CurriculumGraph
-from aigora.curriculum_graph.domain.curriculum_profile import CurriculumProfile
-from aigora.curriculum_graph.domain.edge import Edge
-from aigora.curriculum_graph.domain.enums import EdgeType, MasteryLevel
-from aigora.curriculum_graph.domain.mastery import MasteryCriterion, MasteryScale
-from aigora.curriculum_graph.domain.node import Node
+from aigora.curriculum_graph.domain.entities.curriculum_graph import CurriculumGraph
+from aigora.curriculum_graph.domain.entities.curriculum_profile import CurriculumProfile
+from aigora.curriculum_graph.domain.entities.edge import Edge
+from aigora.curriculum_graph.domain.enums.enums import EdgeType, MasteryLevel
+from aigora.curriculum_graph.domain.value_objects.mastery import MasteryCriterion, MasteryScale
+from aigora.curriculum_graph.domain.entities.node import Node
 
 
 def _make_node(node_id: str = "n1") -> Node:
